hf_space_stage/requirements.txt
# HuggingFace Space — runtime-only deps for the GridSense-AZ Streamlit dashboard.
# Build target: Streamlit SDK on CPU Basic (16 GB).
# Pins track the training environment so the gwnet_v0.pt state_dict loads cleanly.
# NOTE: torch-geometric intentionally omitted — `topology.to_pyg_data` is only
#       used at training time; the dashboard never imports it at runtime.
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.4.1
streamlit==1.38.0
pydeck==0.9.1
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
pyarrow==17.0.0
opendssdirect.py==0.9.4
networkx==3.3
